requirements.txt
asgiref==3.9.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.23
django-debug-toolbar==6.0.0
django-extensions==3.2.3
django-taggit==6.1.0
dotenv==0.9.9
easy-thumbnails==2.10.1
idna==3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
Pillow==9.2.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.2
typing_extensions==4.12.2
urllib3==1.26.6
Werkzeug==3.1.3
